services/agent-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.0
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.23.8
openai==1.40.0
psycopg2-binary==2.9.9
